Elastic constants of Al2O3
These are needed to extract the modulus of Zn from the experimental
results. This function will return the single crystal elasticity
matrix at a chosen pressure and temperature (in K and GPa). The
Young's modulus of an isotropic polycrystal is also returned
(as this is the data we need to use to get E from the ratio of
the amplitudes).