Start of framework to allow three or four foils
And various ways of processing. Use the mode argument
to set the type of analysis. This:

   sine_fit_strain('Zn_07_27tons_26C_300s_007_position_change.txt', ...
        'mode', '-SE')

is supposed to reproduce the older code for four foils, but this
fails to match the amplitude (phase and phase lag is exactly
the same). I think the current version is correct and we
were flipping the reference lengths before.